PLAT-10-313: Form reply follow-up tasks
Support tasks defined after a form reply one, as being dependencies of
the form reply (i.e in the same sub process).

The on/activity dependency does not yet exist to define a task outside
the subprocess.

Also support a debug mode where the bpmn/png are generated (based on the
log level).