PLAT-11085 Send Message Activity (#10)
* PLAT-11085 Send Message Activity:
In this commit, Send message activity is supported when it relies on a previous activity to get stream Id where the message should be sent, and also when the streamId is provided in the workflow YAML file. If both, then the one created by previous activity is used. Integration tests are added as well.

A new field to is added to send-message activity to define the streamId, this will allow us later to support different ways to use to define the recipient like a group of streamIDs.. Output from activities are passed to next ones