Support watch folder to load workflows
A new property has been added to configure a folder where to put
workflows for the bot (the /validate command is not always convenient).

It is watched using the JDK Watchservice so speed may vary based on the
OS (that is why the tests have been disabled, it is slow on Mac).