PLAT-11101: Custom activities
Check the docs/ folder on how activities can be used.

The project has been split in multiple modules so you can implement
custom activities with a limited set of dependencies.

Activity matching is made on the class name (with CamelCase/Kebab-case
mapping).

On the fly the JSON schema is updated to add custom activities so they
are validated (at least their existence and common attributes).