PLAT-11090: Event are passed to activities
Serialized as JSON (handled by Spin module)
Events can be used in activties or scripts (you might have to escape
presentation ML to avoid JSON ser errors).

send-message/to/stream-id is now optional and the current stream (from
the current event) will be used if not set (which is makes it easier to
have a bot that replies to a message).

Better error handling in workflow watcher.