PLAT-11184: Events in the middle of workflows
Before only starting events were supported, we now support events in the
middle of workflows, so one activity could be in a waiting state.

Multiple events are supported too, using the one-of attribute.

As multiple workflows can wait for the same event, Camunda messages have
been replaced with signals. They are more flexible as they can be
broadcasted (whereas messages have to be unique). The event dispatching
changed a bit and we cannot know anymore if a workflow has been started
(or resumed) after an event is sent.