Expose metrics
Start exposing metrics (even custom ones for workflows) with Actuator.
By default it is enabled, with the bot now exposing the 8080 port.

The way metrics are queried might be a bit time consuming (db queries)
but let's keep it simple for now.

Also cleanup the event logger and make tests a bit faster with shorter
timers.