PLAT-11187: Manage room activities
A few new activities have been added:
- edit-room
- add/remove-room-member
- promote/demote-room-owner

Activate a room is covered by edit-room.
edit-room sets the room as output. For the others I thought it was not
worth it and will be possible with the get-room activity.

I had to remove the error handling in CamundaExecutor to let the
workflows fail the unit tests if a script fails for instance. The error
handling will to be revised later (we have a ticket for that).

JSON Schema not provided (we have another ticket for that).