PLAT-11196: Run workflows in parallel
Support workflows listening to the same event to run in //.

When an event is dispatched as a signal, each workflow execution was
starting in the current thread. Which is ok with multiple events because
there are chances the events are dispatched in different threads (we use
the spring boot integration of the BDK which is async by default).

If multiple workflows are reacting to the same event the execution was
not in //. To support that events have been configured to be async
before, i.e the job executor will pick up activities and processes ready
to start instead of running them in the current thread.

This makes the tests a little more complex because we cannot be sure
after sending an event that the workflow is running or not yet. That is
why some tests have been changed to send an event and check the expected
behavior with a polling strategy (using awaitility).

Also fix a bug with multiple events in the middle of workflows that were
not properly connected.