Rename edit-room to update-room
Reasoning being:

As an example, edit-room is a confusing activity name (for someone who does not know/play with our APIs). "Edit" is more to let the user to enter his values. update-room would be more appropriate I think. In general, we have to start referencing the SWADL language and assess all terms are the right one.