PLAT-11240: SWADL validation errors
Errors reported by the JSON schema library we use are a bit complex to
understand and do not provide line numbers.

Here we try to improve that by providing a line number for the most
specific error we find in the report (and the first one) to help the
user fix his workflow.

We also try to provide simple error messages for a few use cases (check
the YamlValidatorTest). Ideally we would like to guide the user to fix
his workflow by suggesting the allowed values for instance but this made
the code even more complex than it is now. So I'd prefer to wait for
feedback with real users to see if want to improve this further.

The JSON schema has been improved a bit, there will be another task to
complete the documentation for all properties.