Validate activity id format
Invalid activity id could cause error such as
org.camunda.bpm.model.xml.ModelValidationException: DOM document is not valid

because ids are used as identifies in BPMN/XML.

We enforce a pattern for ids, it will also help when referencing
variables based on task ids (${} expressions).

Also handle a warning in IDEA about more than one variant for room
activities. Those don't have required properties so the validator is a
bit lost to match the schema with the support of custom activities.