PLAT-11197: User activities
New activities to handle users have been added:
- get-user
- get-users (aka search)
- create-user
- update-user
- add-user-role
- remove-user-role

APIs to manage avatars are not supported (this does not seem super
important) as we would like to deal with attachments first.

Suspend user is not supported by the BDK so it is not implemented.

Get roles seems to be restricted to specific use cases so it has not
been implemented so far ( I expect workflows to use hard coded roles).

We have many APIs to search users, the v3 has been used (just because it
is a v3 honestly I could not figure out which one is better... I'll
check the API usage on Splunk).