PLAT-11195: Support variables everyhwere
In the schema and Java model we cannot use specific types like boolean,
integer if we want to support variables in the SWADL file.

So instead we use string OR boolean/integer and also allow ${} where
patterns are used in the JSON schema.

As the variable resolution happens at runtime we must keep the string
type in the Java model but we provide typed getters.