PLAT-11242: Args for /slash commands
Arguments can be used in message-received event's content. This allows
to match /commands such as: /create-room Room's name for instance.

Mentions, hash and cash tags can be matched too (see the tests for
examples) with a bit of escaping sometime.

The AntPathMatcher from Spring is used, making the implementation super simple and
the usage easy for developers familiar with Spring path matching in
controllers.

To match workflows the problem is that we need to list all the workflows
waiting for a signal and check if they match (as before we could just
send a signal and then Camunda would match it automatically). By
introducing args (i.e regex) we cannot guess the matching workflows so
we have to go over all of them. This might not be efficient if the bot
is running many workflows, maybe in the future we would have to
introduce a flag to disable this syntax and keep event dispatching
efficient (but that would be premature optimization IMHO).