PLAT-11242: Replace Bdk bean with function mappers
Previously we could access some utility functions with ${bdk.text()} for
instance.

It has been changed to a function mapper so we don't even need to add
the bdk prefix, the functions can be used directly.

We also expose methods to access the data associated with a message
(mentions, tags, emojis).