(WIP) PLAT-11130 Support attachments in send-message activity (#23)
Related to [PLAT-11130](https://perzoinc.atlassian.net/browse/PLAT-11130)
In this PR we added support for attachments in send-message activity.
The proposed SWADL defines an attachment with either a content file path and an optional preview file path from file system, or a message id and attachment id if we want to forward an already sent attachment. When only the message id is provided, then all attached files are sent along with the new message. Attachments coming with the slash event are also supported and forwarded.