Fix serialization issue when variables are used
See https://perzoinc.atlassian.net/browse/PLAT-11195?focusedCommentId=787536

If variables are used, then the getter that does the type conversion
could be used before the variable is replaced with its real value (so it
is still a string).

To avoid that different typed getters are provided (and marked as ignore
for Jackson). This way regular string getters are used and we can
use the typed getters in the executors.