PLAT-11197: create-user/create-system-user activities
To better differentiate the 2 different types of users we introduce 2
different activities (and the same for updates). This way the
completion/validation of SWADL is less error prone and better guides the
user with required fields.

The implementation is kept simple with the same executor being reused
and the Java model sharing the same properties for the 2 use cases.