PLAT-11103: Get message activity (#29)
In this PR, two activities have been added: get-message to get a message by id, get-messages to list message in a given stream with pagination.
For the same occasion, we implemented a custom assert upon assertJ for Workflow integration tests. To implement it, we couldn't use HistoryService bean from IntegrationTests class in a static context. The workaround we did was to add a static field as a copy of the bean that is set in a PostContruct method. Thus, the existing one is injected by Spring and the second one is instantiated post contruct and get used in custom assertion class. This can be improved in a later commit.