PLAT-11209: JSON schema refinement (#38)
Added description to some properties fields using x-intellij-html-description to display description and links in Intellij popups. However, nothing was provided for VSCode and other editors.
Removed version field from Workflow as we only support v1 for the moment and replaced workflows name field by id and made it optional. If no id is provided, the yaml filename is used. Otherwise, the id is randomly generated with a fixed prefix. Some fields have been factored to optimise the json schema.
For the same occasion, we refactored all fields previously named uids to userIds for consistency.
Finally, we updated some properties definitions with enum types in order to let IntelliJ suggest replacement for a better autocompletion.