PLAT-11286 PLAT-11287 Error management (#43)
In this commit, we added an exception UniqueIdViolationException to be thrown when a workflow has activities with the same id. The exception gives more context with the duplicated ids and the workflow id which can be helpful in case multiple workflows are being deployed.

In the same way, NoStartingEventException message is changed to show the workflow id.
Lombok sneakythrows annotations were removed from the project. They were replaced by a try/catch as the code blocks are not business and the thrown exception should be enough in the logs.

We added InvalidActivityException and ActivityNotFoundException that are thrown when we try to build an invalid gateway in the bpmn or reference an unknown activity id when using form-reply or activity-completed. 

Loops relies on activity-completed property like Conditional branching. The main difference is that an activity in a Loop references another one that comes later in the swadl while in Conditional branching, activities should reference an activity already defined in swadl. This prevents un from throwing an exception when a conditional branching activity references another one that comes later.

Some exceptions were updated with the workflow or the activity id for more context.

Fixed some autocompletion values that were not working. For user-ids-inner, which is defined as a list of user-id, we added an empty enum as a workaround to have autocompletion on values we suggest for user-id.

Added Nullable annotation for all fields that can be null regarding the swadl-schema.
Removed name field from activities as it is not used anywhere. Removed BaseEvent class for the same reason.
Updated the swadl-schema to put some required fields. get-streams/get-user-streams/get-users should be reviewed in next commit.
Updated doc and examples to make swadl files valid regarding all changes we recently did

Finally, workflow-bot-app/workflows was removed and the README.txt is moved to ./workflows, as the default configuration references the root folder