Corrected get-users and create-room activities and improved test coverage
In this commit, we corrected two activities:
- Made get-users activity to require one of the following fields: user-ids, emails or usernames. One and only one of them should be set. 
- Added missing attributes to create-room activity. It was the first activity we implemented at the beginning and we only supported few attributes to show it works. After that, we forgot to add all attributes.
- Changed some fields naming for coherence in update-room and create-room.

For the same occasion, we improved test coverage:
Introduced dependencies property in swadl-schema to check the presence of both skip and limit fields or none of them in get-user-streams. Before, we were checking this error case in the executor and we throw an exception. The code for that was removed in this commit and replaced by the check at swadl-schema level.
Since we added required fields in swadl-schema for get-users activity, the code in executor that was checking if neither usernames, userids and emails are set was removed and is now verified at validation level.
Added some integration test scenarios for send-message activity, stream and users.

Fix a gradle check issue to avoid potential NPE.

Corrected documentation link for get-users activity.