CI using Github actions
One job to run unit tests/static code analysis on PRs.

One job to deploy SNAPSHOTs of the JARs on Maven Central.
This way the generator will be able to use those.

We use SNAPSHOTs only for a short update/feedback loop for now at some
point we might move to released versions.