PLAT-11326: Expose workflow trigger endpoint (#47)
A POST endpoint is exposed to trigger workflows. It takes a workflow id, token in the header and a map in the body request. A new event RequestReceivedEvent is added and dispatched when an HTTP request arrives.
The controller executes the workflow using a new method added in WorkflowEngine. This latter verifies the workflow id. If not present, then a 404 is returned. Otherwise, it checks then the token and 403 is returned if it doesn't match the one provided in the request. 

The token is stored in the workflow as a deployment resource. Token management is simple for the moment and no claims/entitlements are supported for now.
We suppose that a workflow can only have one token to be used by different activities and we do not support different tokens for different activities within the same workflow.

A webMvc based test is added for the new api and one integration test is added for this new event. More tests will be added in next commit.

For the same occasion, the app is configured to prefix all apis with /wdk and put actuator endpoint port to 8081.

Removed the PresentationMLException from execute() method, we now catch it and throw a RuntimeException which is not exposed in the controller.