PLAT-11148: Error handling
Retries have been disabled at Camunda level, so a failing activity/task
only fails one and is not retried (was 3 times by default).

It is up to the workflow write to handle errors (the BDK also retries
when calling API, that can solve intermetting network issues for
instance).

An on.activity-failed event has been added to support that, it is
similar to the activity-expired or activity-completed event.

Using one-of, an error handler activity can be set for multiple
activities. Supporting a generic/catch all error handler was not easy to
implement and raise questions for complex flows or forms (what is the
scope of the error handling then?) so it is not supported for now.

Script task (execute-script) are a bit different so a bit of a hack had
to be used to make sure all exceptions are also caught by
activity-failed events.