PLAT-11280: Add get-attachment activity
A new activity has been added to get one attachment given message and attachment ids. Both attributes are required for this activity.
The activity downloads the attachment, decodes it and stores it in file system at ./workflows folder. Thus, the activity does not have the same output as the api. While this latter returns the attachment content as response, the activity stores it locally and returns the path.
In order to get the ./workflows folder path, a new functional interface ResourceFolderProvider was added since ResourceProvider could not have a second abstract method. The good practice is to keep only one abstract method in a functional interface.
The integration test with the working scenario results in file created in resources, at the end of the test we check that the file is present and we delete it. When it is not deleted, only a log is printed and this does not impact the test result passed/failed.
The reference documentation has been updated as well.

Finally, there is some duplicated code between GetAttachment and SendMessage part that handles attachments, this can be refactored in a next commit. The same for tests as well