PLAT-11401: Support variables usage for list/map attributes
In activity attributes with type list or map it was not possible to use
variable references (${}).

How are variables handled in workflows?

In SWADL you use them with ${}
When SWADL is deserialized to the Java model, strings were used
The Java model activity is serialized to JSON and stored in BPMN
When the BPMN is executed, expression (including variables) are
resolved and real values are replaced.
When the CamundaExecutor runs, the JSON activity is deserialised and the
Java mode rebuilt to be used in the Activity executor.

We now use a custom Jackson deserializer to be able to store in the Java
model either a resolved value or a variable reference (i.e the Variable
class).

We then use a custom TypeConverter during expression resolution (juel)
in Camunda to store complex objects (list,map) as escaped JSON.

When we deserialize to rebuild the Java model in CamundaExecutor another
custom deserializer is used to parse this escaped JSON and preserve the
complex objects.

Activities that have attributes other than strings should wrap them in
Variable<?>. Given JSON only support a generic number type, it is
difficult to differentiate a float/integer/long so it is up to the
activity developer to handle it.

The schema has been changed to allow to use variables or lists or maps
for the attributes that support it.