Made use of the existing interface for storing attachment files.
In the previous commit, we had 2 interfaces one to get resources and another one to store them. In this commit, we merged them in the same interface.
Also, we were storing files in filesystem during test executions and delete them at the end, we changed the interface used in tests in order to mock the file creation and storing.
Fixed some typos and formating issues