Store attachments in a process specific folder.
When an attachment is downloaded, we suffix its name with the activity id which is unique to the instance under execution, and we store it in a folder named with the process id. Thus, if a workflow is executed twice at the same time, then it has two different processes and they don't overwrite each others resources.
If in the same process we have more than one activity handling an attachment file, the files will not have the same suffix and they stay unique to each activity.
Code coverage is under 0.9 required by the checks as we do not test the folder/file creation and writing as we dont want to use the filesystem during tests execution