Support persistent database.
Previously, we were only supporting in-memory h2 database. Hence, workflows state and data was lost when the app restarts.
The datasource configuration was done in a bean as part of camunda engine configuration class. In this commit, we moved that configuration to application.yaml file to allow users to switch it to a persistent database.
The persistent database is an H2 file based one, it can be enabled by changing the database url in application.yaml file.

Deployment documentation was changed to explain that.

For the same occasion, Camunda bpmn builder was changed to not add suffix to workflows names/ids. Actually, using the persistent database, when we deploy a workflow that has already been deployed, if they have the same id, then only the second one is used for new executions. The pending executions use the old one when they are relaesed.
To support that, we had to change the builder to not add suffix. Thus, workflows can be versionned when a new one is deployed.