Provided yaml filename to deploy() as default workflow id if no one is provided
The method deploy() was only taking the workflow as params, now we added the yaml filename to use it as id if no id is provided.
Enabled the persistent database by default in integration tests to be able to add a scenario where we deploy 2 workflows with the same id and check that only the second one is used.
We might need to add another one to verify that suspended processes executed before deploying the second workflow keep using the first one.

Updated deployment doc with relevant info about the JDBC driver and removed useless quotes in configuration.