Support H2 persistent database (#18)
Previously, we were only supporting in-memory h2 database. Hence, workflows state and data was lost when the app restarts.
The datasource configuration was done in a bean as part of camunda engine configuration class. In this commit, we moved that configuration to application.yaml file to allow users to switch it to a persistent database.
The persistent database is an H2 file based one, it can be enabled by changing the database url in application.yaml file.

Deployment documentation was changed to explain that.

For the same occasion, the workflow id is set to be required. Actually, using the persistent database, when we deploy a workflow that has already been deployed, if they have the same id, then only the second one is used for new executions. The pending executions use the old one when they are released.
To support that, we had to make the workflow id required. Otherwise, no versioning is possible when a workflow does not have an id. Thus, workflows can be versionned when a new one is deployed.