PLAT-11371: Added curl activity.
In this PR, we added a new activity send-request that can be used to perform HTTP requests. Its outputs are the request status and the response body if it succeed.
A dependency id added to include BDK http jeresey2.
We reuse the same bot configuration for api client, so no additional configuration is needed. This way, the proxy and other settings already set can be used.
The api client is supposed to be used to target different apis and not necessarily the pod. Hence, the host url changes in every request. Thats why the apiClient is reinitialized with a new basepath for each execution. The response body is stored in outputs and we dont expect to have a big amount of data returned by the api. In this case, the disk could be full.

The reference documetation and integration tests have been updated for this occasion.