Fixed integration tests. > > When added the integration test for ExecuteRequest activity, we had to inject a mocked bean of SpringBdkGateway in the parent integrationTest class. This broke all the tests. In fact, we inject mocked beans of BDK services (messageService, streamService..) in this parent class. When we injected a mocked bean of SpringBdkGateway, which contains bdk services itself, then this latter ones are used in tests instead of the ones mocked/injected one by one. Thats why tests were broken.
Why couldnt we simply inject the apiClient bean in the test parent class, like othere bdk services insteaed of injecting the whole SpringBdkGateway? It was not possible since the apiClient bean cannot be exposed in the gateway. Actually, we instantiate ApiClientFactory with bdk configuration bean, and we access the apiClient with a getter method that usees the apiClientFactory to create a new instaance of it. This happens since the host url is not fix for the apiClient.

The solution we came with in this commit, is that we keep injecting/mocking bdk services one by one in test parent class beside SpringBdkGateway, but in the test setup, we force returning the right services when the mocked SpringBdkGateway is called. Another solution would be to make those beans priority by using a Spring annotation.