Made use of ServiceLookup to load apiClient.
Since wdk relies on springboot, webclient should be used instead of jersey2. For this, we made use of ServiceLookup to load the apiClient.
We got rid of proxy and trust store configuration for the following reason: it would be better to have a wdk sepcific configuration to set trustores and proxies. With what is provided for the moment, we cannot configure the apiclient to target different apis that use each one different certs/keys and proxies.
Added an example where we call a public api to dispplay exchange rates and sends EUR and USD rates in message. This was added to show how we can process the api request response in the swadl.