Storing request response error message as JSON format in outputs.
When a request fails, we store the error message returned by getResponseBody() as data in outputs. ApiException.getResponseBody() returns a string wich prevent processing the response in subsequent SWADL activities. We map the string to a Map before storing it in outputs.
Added unit tests with a script activity that checks outputs values. In case the actual value is not expected, the script activity makes the workflow fail which makes the test fail as well.