PLAT-11371: Added curl activity (#24)
In this PR, we added a new activity send-request that can be used to perform HTTP requests. The outputs contain the response status and body. In case the request have failed, the error message from response body is stored in the body as well. We dont expect to have a big amount of data returned by the api. In this case, the disk could be full.

The api client is supposed to be used to target different apis and not necessarily the pod. Hence, the host url changes in every request.

The api client is not configured with proxy and trust store so far. Actually, WDK uses the bot configuration. Hence, we cannot configure the api client for different apis that have different proxis or require different certs/keys. A good solution would be to add WDK specific configuration and set proxi and cert config for apis in the SWADL when writing the request activity.

The reference documentation and integration tests have been updated for this occasion.