PLAT-11328: Prevent stale workflows from running forever
Every event starting an activity can define a timeout property. It will not be executed if it timeouts. This prevent stale workflows from running forever.
It is the responsibility of the workflow developer to set the timeout for activities.
Previously, the default timeout was set to 24h, this has been changed in this PR and no default value is set. Actually, the bpmn builder checks for the timeout property to set duration in the bpmn. The default value was removed so as the builder finds a null value and does not handle it. Otherwise, it would create a subprocess for each activity with duration of 24h.