PLAT-11328: Prevent stale workflows from running forever.
Every event starting an activity can define a timeout property. It will not be executed if it timeouts. This prevent stale workflows from running forever.
It is the responsibility of the workflow developer to set the timeout for activities.
We kept default timeout of 24H for forms. Instead of having it as default value of the property in the pojo, we set it at CamundaBpmnBuilder in order not to be applied for other type of activities. This means no default timeout for activities except the forms.
Otherwise, the bpmn builder would create a subprocess for every activity with duration of 24H if it is kept as default.