Made test faster and put form activity default timeout a constant.
In the timeout test, we reduced the timeout in SWADL to small value in order to make the test faster as the thread sleeps less time.
Form replied events have a default timeout of 24H. This is a good candidate to be put as WDK config but as we do not have such config for the moment, we put it as a private final constant.