Bumped version to 0.0.6-SNAPSHOT
An artifact with previous version 0.0.5 was already published on Nexus during local testing. We would like to check if the publish workflow action works correctly when this PR is merged on master. To check that, we should see a new version published on Nexus. to be able to do so, we bumped the version.