[CICD] PLAT-11425: Release process for publish snapshots and closed releases (#54)
In this commit, we added publishAndCloseRepository task to the command executed in release.yaml file that is executed when a tag/release is created on Github.
This task comes as part of nexus-staging plugin and this latter should be defined in the root level. That is why we added a build.gradle file at root of project and removed duplicated properties from workflow-bot.java-conventions.gradle. Now, the group, version and pom definition is done at root build.gradle.

We disabled the task (jar) in order to prevent publishing plain jar that is not used.
Added bootJar to publish the fat jar used by Symphony generator.