Improvement in execute-request activity and documentation.
An API can return a simple String as exception message (not key/value format). Since this is considered as a valid JSON, we have to support it. In this case, we create a key (message) and put the string inside before setting it in activity outputs.

Documentation updated to mention that timeout is set to 24H by default for form-replied event, and no default for other activities.