Support lists in request headers and made HTTP method optional with GET default value.
The method property in execute-request activity is optional now with a default value of GET, given that it the most used in HTTP clients. Thus, one less field to set in the activity.
For the same occasion, we changed the headers to support lists. When a list if passed in the activity headers, we join the list items in a string separated with a comma. When converting a list to a string, this latter has opening and closing brackets which should not be present there. We replace them with an empty string. If the item value itself contains a bracket, it will be removed as well. We did this considering that such characters are not expected in headers values. Otherwise, a different implementation should be done to build the headers string by putting the items one by one using a map function for example.