Fix variable deserialization for lists
Variables that are lists were supported, however list of variables were
not. We now support both by declaring the activity attributes as
Variables of List of Variables.

Ser/Deser has been updated to reflect that. We do not consider this a
perfect solution for now as it makes the activity writing a bit
cumbersome. Also maps are not fully supported here.

Fixes #61 PLAT-11547