Support different HTTP methods in execute-request activity.
In this commit, we support different HTTP methods and set GET as the default one.
We get ride of the bdk http client and used apache client5 instead in order to configure easily and without requiring any change in bdk.
execute-request returns now the response body as a string in addition to the response status code. In next commit, we will add a utility method to read the output string as a json and process its content.

The http client was only required by requestExecutor and is injected as a bean. This requires requestExecutor to be a bean as well. It is annotated as a component. Hence, every executor class can be a bean or not. In CamundaExecutor, we load executors as beans by using Spring application context. In case the bean is not found, because the executor is not a bean for example, then the exception is catched and a new instance is created instead of loading the bean.