Added utility function to process response outputs as JSON.
In this commit, we added a public static method in BDKFunctions that takes the response body which is stored as a string and returns it as a JSON. It can be used in SWADL file.
The body of requests is set as a string when the content type is application/json. Otherwise when it is multipart form data, it is set using a form. This still need an improvement and testing.