Updated json utility function and request body type.
We were not able to process JsonNode in SWADL so we changed json utility function to return an Object to be more generic. Properties are returned in their actual types: boolean, String, Double.. We can now iterate deeply in an api response in SWADL level.
The body of requests has been changed to support both object and string. In case of multipart-data-form content type, we want to simply cast the params to a map at the httpclient level. To be able to do so, the body should be provided as an object in SWADL level. For the other types, we need to provide the body as a string in httpClient level along with the content type. This required to provide it as a String in the SWADL level in order to avoid further casts.