PLAT-11535: Timeout support without using subprocess.
To support timeout in different activities, an implementation was done based on subprocesses but it turned out that it is not the correct implementation in terms of flow control as subprocesses can be rexecuted multiple times.
In this commit, we implement it in the main process. For this occasion, we introduced EventWithTimeout than represents an event that can have a timeout, like this in the java model we do not allow to have timeouts in one-of event and it can only be set at (on) property level.