PLAT-11535: Timeout support without using subprocess (#65)
To support timeout in different activities, an implementation was done based on subprocesses but it turned out that it is not the correct implementation in terms of flow control as subprocesses can be rexecuted multiple times.
In this commit, we implement it in the main process. For this occasion, we introduced EventWithTimeout than represents an event that can have a timeout, like this in the java model we do not allow to have timeouts in one-of event and it can only be set at (on) property level.

For the same occasion, we added the support of on:acitivty-expired for all activities since they can have timeout now. Previously, an activity could not be defined to be executed on multiple activities expirations, only one could be used.
When an activity has on expiration property, so the list of activities to be expired are processed.

It was difficult to add the activity-expired for different activities to what was implemented previously for form expiration. This latter was implemented as a subprocess while for the other activities we use the main process. At a point when building the bpmn, we were considering that all bpmn nodes with timeout property were subprocesses and have to be finished, which does not work when the bpmn node is not a subprocess and this makes the workflow building fails. A new map to store only subprocesses is added and used when we finish them.