Fix javadoc generation
Javadoc is currently failing and preventing deploying on Maven Central.

Fixed the issues (<p> tags) and made it run along the check tasks so we
avoid merging incorrect javadoc.

Also disabled Javadoc warnings as we don't need to address them
(sometimes we are better off with no @return/@throws if we have nothing
interesting to comment).