PLAT-11536 Fixed multipart content type. (#66)
When sending a multipart/form request, the content type should include the boundary. However, this is no something the user can set in SWADL.
In this commit, we changed the way we handle this content type request and we made use of MultipartEntityBuilder that sets the boundary along with the content type in body.
When the content type is provided in the header than it overrides the one set by MultipartEntityBuilder. For this reason, when we process a multipart/form request, we remove the content type from headers if the user has provided any in order to make use of the one set with boundary.

Updated reference doc with some UtilityFunctions