PLAT-11517: acitivity outputs get overridden
Some activities have more than one output variable to store in the context. Outputs are stored in an outer map having one item key called outputs, and the value is an inner map storing the actual variables. The method setOutputVariable() that was used for that was handling one variable at a time. When it is called a second time to store the second variable, then it was creating a new outer map with outputs key before setting the actual variable which overrides the previosu output in the context. In this commit, we refactored that with a method that takes a map of outputs and store them in one shot.
An output might be null, so we use a Hashmap and update it by putting outputs one bye one instead of doing it in one line.
Some integration tests were updated to verify multiple outputs are stored in the context