Improvement of execute-request activity.
Execute-request activity supports now a body that can be either put in json string or object. HttpClient is able to do the required conversion depending on the body instance type.
This change makes the workflow writing easier.
Body field in Response object is now an Object to support both string and Map.