Fixed Deserialization bug on immutable collection storing workflows outputs.
In #69, we created the outputs outter map using Map.of() that creates an immutable map. This map was supposed to be read only as a new instance is created at the end of every activity, and we never updates it. To update an output we create a new instance of the map.
This introduced a new bug in the ProcessEngine. When an event comes to the workflow, the ProcessEngine deserializes the context outputs, and this fails when the map storing context outputs is immutable. The error could not be found on basic workflows that have one event to start it, as at the beggining no immutable map is created in the context yet. However, with events in the middle of the workflows, an immutable map is already found in the context and outputs deserialization fails with org.camunda.bpm.engine.ProcessEngineException.
This has been fixed by creating the map using new hashmap().